opti: Using full potential of mirror matrix
Seams like immintrin load instructions are not to
be used in a concurrent way. Using a simple
memcpy() enables the full potential of sequential
reads on a C table.
With the mirrored table in each matrix there are
less cache misses on matB during comput_case_simd()
which improves a lot the performance (x2.35).